requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
chardet==4.0.0
click==7.1.2
digi-xbee==1.3.0
future==0.18.2
idna==3.1
importlib-metadata==3.7.2
itsdangerous==1.1.0
MarkupSafe==1.1.1
multidict==5.1.0
pefile==2019.4.18
pyserial==3.5
python-dateutil==2.8.1
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2020.5
pywin32-ctypes==0.2.0
retrying==1.3.3
six==1.15.0
srp==1.0.16
typing-extensions==3.7.4.3
websockets==8.1
yarl==1.6.3
zipp==3.4.1
